Refactor analytics and exports modules to utilize centralized database utility functions
