+++ api/routers/analytics.py
@@ -1,17 +1,14 @@
 """
-This module provides analytics endpoints for retrieving and visualizing
-expense data. It includes routes to generate visualizations for expenses
-from a specified number of days.
+This module provides analytics endpoints for retrieving and visualizing expense data.
 """
 
 import datetime
 from typing import Optional
 
-from bson import ObjectId
 from fastapi import APIRouter, Header, HTTPException, Response
-from motor.motor_asyncio import AsyncIOMotorClient
 
 from api.utils.auth import verify_token
+from api.utils.db import fetch_data
 from api.utils.plots import (
     create_budget_vs_actual,
     create_category_bar,
@@ -19,47 +16,8 @@ from api.utils.plots import (
     create_expense_bar,
     create_monthly_line,
 )
-from config.config import MONGO_URI
-
-# MongoDB setup
-client: AsyncIOMotorClient = AsyncIOMotorClient(MONGO_URI)
-db = client.mmdb
-users_collection = db.users
-expenses_collection = db.expenses
-accounts_collection = db.accounts
 
 router = APIRouter(prefix="/analytics", tags=["Analytics"])
-
-
-# Utility function to fetch data
-async def fetch_data(
-    user_id: str, from_date: Optional[datetime.date], to_date: Optional[datetime.date]
-):
-    """Fetch data from the database based on user ID and date range."""
-    if from_date and to_date and from_date > to_date:
-        raise HTTPException(
-            status_code=422,
-            detail="Invalid date range: 'from_date' must be before 'to_date'",
-        )
-
-    from_dt = (
-        datetime.datetime.combine(from_date, datetime.time.min) if from_date else None
-    )
-    to_dt = datetime.datetime.combine(to_date, datetime.time.max) if to_date else None
-
-    query = {"user_id": user_id}
-    if from_dt and to_dt:
-        query["date"] = {"$gte": from_dt, "$lte": to_dt}  # type: ignore
-    elif from_dt:
-        query["date"] = {"$gte": from_dt}  # type: ignore
-    elif to_dt:
-        query["date"] = {"$lte": to_dt}  # type: ignore
-
-    expenses = await expenses_collection.find(query).to_list(1000)
-    accounts = await accounts_collection.find({"user_id": user_id}).to_list(100)
-    user = await users_collection.find_one({"_id": ObjectId(user_id)})
-
-    return expenses, accounts, user
 
 
 @router.get("/expense/bar")
@@ -162,24 +120,8 @@ def prorate_budget(
     last_expense_date: Optional[datetime.date],
 ) -> float:
     """Prorate the budget based on the date range."""
-    if from_date and to_date:
-        days_in_range = (to_date - from_date).days + 1
-    elif from_date:
-        days_in_range = (
-            (last_expense_date - from_date).days + 1
-            if last_expense_date
-            else (datetime.date.today() - from_date).days + 1
-        )
-    elif to_date:
-        days_in_range = (
-            (to_date - first_expense_date).days + 1
-            if first_expense_date
-            else (to_date - datetime.date(1970, 1, 1)).days + 1
-        )
-    else:
-        days_in_range = 30  # Default to 30 days if no date range is provided
-
-    print(f"Prorating budget: {budget}, Days in range: {days_in_range}")
+    from api.utils.db import calculate_days_in_range
+    days_in_range = calculate_days_in_range(from_date, to_date, first_expense_date, last_expense_date)
     return (budget / 30) * days_in_range
